+++ templates/install/class_for.rb
@@ -1,3 +1,0 @@
-  def class_for(name, from: identifier)
-    "c-\#{from}-\#{name}"
-  end