+++ templates/install/index.js
@@ -1,2 +1,0 @@
-const context = require.context(".", true, /index.js$/)
-context.keys().forEach(context);